requirements.txt
setuptools
lingua-language-detector
unidecode
pykakasi
pinyin
wheel
tempora==1.14.0
jaraco.functools==2.0
six
Flask
flask-sock
pillow
qrcode
pygame; sys_platform != 'darwin'
pygame==2.1.3; sys_platform == 'darwin'
futures
psutil 
unidecode
requests
arabic-reshaper
python-bidi
flask-paginate==2021.10.29
flask==2.3.1
numpy
torch>=1.5.1
tqdm>=4.30
librosa
pip
soundfile
spec_utils
